fix issues with 64-bit printing
1) int64_t is not always "long long".
on my machine and others it is "long".
there is a portable way to give int64_t
to printf(), that is the PRId64 macro
provided by inttypes.h, which is POSIX
standard. Windows doesn't have it though.

2) some code was casting off_t to (long long)
and then printing it using INT64_T_FMT.
we know it is long long now, so just use %lld.