require CMake version 3.0.1, fix default install
many great features were added in version 3,
and @LoicMarechal says its fine to raise the requirement.

After lots of Googling, it seems the nicest way
to have a default CMAKE_INSTALL_PREFIX is described
here:

https://cmake.org/cmake/help/latest/variable/CMAKE_INSTALL_PREFIX_INITIALIZED_TO_DEFAULT.html

I have tested that things install to /opt/libMeshb
if the variable is not defined, and if it is defined
then it goes to the user-specified location.